Added GAMES_PATH option to .env Updated portal
