+++ .env.example
@@ -1,1 +1,2 @@
 PORT=8000
+GAMES_PATH="./games"
