+++ public/assets/index.css
@@ -4,18 +4,25 @@ body {
 }
 
 .Games {
-    margin: 30px;
+    margin: 5px;
     text-align: center;
 }
 
 .Game {
-    margin: 15px;
+    margin: 5px;
+    padding: 15px;
     display: inline-block;
     text-align: center;
     cursor: pointer;
+    border: 5px solid white;
+    border-radius: 30px;
 }
 
-.Game[disabled] {
+.Game[disabled=false]:hover {
+    border: 5px solid black;
+}
+
+.Game[disabled=true] {
     opacity: 0.3;
     cursor: not-allowed;
 }
