+++ public/index.html
@@ -10,10 +10,10 @@
       integrity="sha256-pvPw+upLPUjgMXY0G+8O0xUf+/Im1MZjXxxgOcBQBXU="
       crossorigin="anonymous"
     ></script>
-    <script src="index.js" type="module" defer></script>
-    <link rel="stylesheet" href="index.css" />
+    <script src="/assets/index.js" type="module" defer></script>
+    <link rel="stylesheet" href="/assets/index.css" />
   </head>
   <body>
-    <div id="app">
+    <div id="app"></div>
   </body>
 </html>
