+++ src/configs/games.json
@@ -1,49 +1,49 @@
 [
     {
         "name": "Treasure Hunt",
-        "path": "./games/treasurehunt",
+        "path": "/treasurehunt",
         "image": "treasurehunt.jpg",
         "repository": "https://github.com/rodrigocoeio/treasurehunt.git",
         "port": 89
     },
     {
         "name": "Guess What",
-        "path": "./games/guesswhat",
+        "path": "/guesswhat",
         "image": "guesswhat.jpg",
         "repository": "https://github.com/rodrigocoeio/guesswhat.git",
         "port": 88
     },
     {
         "name": "Flash Learning",
-        "path": "./games/flashlearning",
+        "path": "/flashlearning",
         "image": "flashlearning.png",
         "repository": "https://github.com/rodrigocoeio/flashlearning.git",
         "port": 90
     },
     {
         "name": "Hide & Seek",
-        "path": "./games/hideandseek",
+        "path": "/hideandseek",
         "image": "hideandseek.png",
         "repository": "https://github.com/rodrigocoeio/hideandseek.git",
         "port": 91
     },
     {
         "name": "What Am I",
-        "path": "./games/whatami",
+        "path": "/whatami",
         "image": "whatami.png",
         "repository": "https://github.com/rodrigocoeio/whatami.git",
         "port": 92
     },
     {
         "name": "Secret Words",
-        "path": "./games/secretwords",
+        "path": "/secretwords",
         "image": "secretwords.png",
         "repository": "https://github.com/rodrigocoeio/secretwords.git",
         "port": 93
     },
     {
         "name": "Grammar Structure",
-        "path": "./games/grammarstructure",
+        "path": "/grammarstructure",
         "image": "grammarstructure.png",
         "repository": "https://github.com/rodrigocoeio/grammarstructure.git",
         "port": 94
